chore(release): 0.21.0

+++ integ/scripts/bash/cleanup.sh
@@ -13,9 +13,15 @@ shopt -s globstar
 INTEG_ROOT="$(pwd)"
 
 for COMPONENT in **/cdk.json; do
-       COMPONENT_ROOT="$(dirname "$COMPONENT")"
-       rm -f "${COMPONENT_ROOT}/cdk.context.json"
-       rm -rf "${COMPONENT_ROOT}/cdk.out"
+    # In case the yarn install was done inside this integ package, there are some example cdk.json files in the aws-cdk
+    # package we want to avoid.
+    if [[ $COMPONENT == *"node_modules"* ]]; then
+        continue
+    fi
+
+    COMPONENT_ROOT="$(dirname "$COMPONENT")"
+    rm -f "${COMPONENT_ROOT}/cdk.context.json"
+    rm -rf "${COMPONENT_ROOT}/cdk.out"
 done
 
 rm -rf "$INTEG_ROOT/node_modules"
